+++ src/page/MyPage/MyInfoSaved.tsx
@@ -1,11 +1,7 @@
 import { useState, useEffect, useRef } from 'react';
-import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-import { RootState } from '../../store/store';
 import axios from 'axios';
-import { Link } from 'react-router-dom';
-import formatDate from '../../utils/formatDate';
-import { BsEyeFill, BsFillSuitHeartFill } from 'react-icons/bs';
+import MySaved from './MySaved';
 
 interface Save {
   id: number;
@@ -25,10 +21,8 @@ export default function MyInfoSaved() {
   const [isLoading, setIsLoading] = useState(false); // 로딩 상태
   const [isEndPage, setIsEndPage] = useState(false); // 페이지 끝 상태
   const targetRef = useRef<HTMLDivElement | null>(null);
-
-  const token = useSelector((state: RootState) => state.token.token);
+  const observerRef = useRef<IntersectionObserver | null>(null);
   const [page, setPage] = useState(0);
-  const [pageSize, setPageSize] = useState(6);
 
   // useEffect(() => {
   //   fetch('/address/api/saved')
@@ -40,33 +34,32 @@ export default function MyInfoSaved() {
   // }, []);
 
   useEffect(() => {
-    const Access_token = localStorage.getItem('token');
-    const fetchUserPost = async () => {
+    const fetchUserSaved = async () => {
+      if (isEndPage || 0 === page) return;
+      setIsLoading(true);
+      const Access_token = localStorage.getItem('token');
       try {
-        const response = await axios.get(`/address/api/users/planners/userSave?${page}&pageSize=6`, {
+        const response = await axios.get(`/address/api/users/planners/userSave?page=${page}&pageSize=${PAGE_SIZE}`, {
           headers: {
             Authorization: `Bearer ${Access_token}`,
           },
         });
-
-        if (response.data) {
-          const { content } = response.data;
-
-          if (content > PAGE_SIZE) {
-            setIsEndPage(true);
-          }
-          setSavedData(content);
+        setIsLoading(false);
+        const newData = response.data.content;
+        if (newData.length === 0) {
+          setIsEndPage(true);
         } else {
-          console.log(response);
-          alert('사용자 정보가 없습니다 로그인확인해주세요');
+          const threeItems = newData.slice(0, PAGE_SIZE);
+          setSavedData(prevData => [...prevData, ...threeItems]);
         }
       } catch (error) {
-        console.error('사용자 정보 가져오기 오류 확인바람(저장함):', error);
+        setIsLoading(false);
+        console.error('데이터 요청 실패:', error);
       }
     };
-    console.log('저장함', savedData);
-    fetchUserPost();
-  }, [token, page]);
+
+    fetchUserSaved();
+  }, [page, isEndPage]);
 
   useEffect(() => {
     if (savedData.length >= 2) {
@@ -80,88 +73,45 @@ export default function MyInfoSaved() {
     const observer = new IntersectionObserver(handleIntersection, {
       root: null,
       rootMargin: '0px',
-      threshold: 1,
+      threshold: 0.5,
     });
-
+    observerRef.current = observer; // useRef로 observer 저장
     if (targetRef.current) {
       observer.observe(targetRef.current);
     }
-
-    function handleIntersection(entries: IntersectionObserverEntry[]) {
-      entries.forEach(entry => {
-        if (entry.isIntersecting) {
-          const nextPage = page + 1;
-
-          //  setPageSize(prevPageSize => prevPageSize + PAGE_SIZE);
-
-          setIsLoading(true);
-
-          const Access_token = localStorage.getItem('token');
-
-          axios
-            .get(`/address/api/users/planners/userSave?page=${nextPage}&pageSize=3`, {
-              headers: {
-                Authorization: `Bearer ${Access_token}`,
-              },
-            })
-            .then(response => {
-              setIsLoading(false);
-              const newData = response.data.content;
-              const ThreeItems = newData.slice(0, PAGE_SIZE);
-              setSavedData(prevData => [...prevData, ...ThreeItems]);
-              setPage(nextPage);
-            })
-            .catch(error => console.error('데이터 요청 실패:', error));
-        }
-      });
-    }
-
     return () => {
       if (targetRef.current) {
         observer.unobserve(targetRef.current);
       }
     };
-  }, []);
+  }, [page, isLoading]);
+
+  function handleIntersection(entries: IntersectionObserverEntry[]) {
+    entries.forEach(entry => {
+      if (entry.isIntersecting && !isLoading) {
+        setPage(prevPage => {
+          if (!isLoading) {
+            return prevPage + 1;
+          }
+          return prevPage;
+        });
+        if (targetRef.current && observerRef.current) {
+          observerRef.current.unobserve(targetRef.current);
+        }
+      }
+    });
+  }
 
   return (
     <SavedContainer className={containerClassName}>
       {savedData.map(item => (
-        <MySaved key={item.id} savedData={item} />
+        <MySaved key={item.plannerId} savedData={item} />
       ))}
       {!isEndPage && <ObserverTarget ref={targetRef} />}
       {isLoading && <LoadingMessage>로딩 중...</LoadingMessage>}
     </SavedContainer>
   );
 }
-
-const MySaved = ({ savedData }: { savedData: Save }) => {
-  return (
-    <BoxWrap>
-      <Link to={`/page/${savedData.plannerId}`} key={savedData.plannerId}>
-        <Box>
-          <ImgDiv>
-            <TextImg src={savedData.thumbnail} alt="첫번째 이미지" />
-          </ImgDiv>
-          <Info>
-            <TopContainer>
-              <IconWithCount>
-                <Heart />
-                <Count>{savedData.likeCount}</Count>
-              </IconWithCount>
-              <IconWithCount>
-                <LookUp />
-                <Count>{savedData.views}</Count>
-              </IconWithCount>
-            </TopContainer>
-            <TitleText>{savedData.title}</TitleText>
-            <ScheduleeText>{savedData.title}</ScheduleeText>
-            <DateText>{formatDate(savedData.createAt)}</DateText>
-          </Info>
-        </Box>
-      </Link>
-    </BoxWrap>
-  );
-};
 
 const SavedContainer = styled.div`
   display: flex;
@@ -187,99 +137,4 @@ const ObserverTarget = styled.div`
 const LoadingMessage = styled.p`
   font-size: 36px;
   font-weight: bold;
-`;
-
-const BoxWrap = styled.div`
-  margin-right: 30px;
-  margin-bottom: 20px;
-
-  &:nth-child(3n) {
-    margin-right: 0;
-  }
-
-  /* @media screen {
-    &:nth-child(2n) {
-      margin-right: 0;
-    }
-  } */
-`;
-
-const Box = styled.div`
-  position: relative;
-  width: 100%;
-  height: auto;
-  cursor: pointer;
-`;
-
-const ImgDiv = styled.div`
-  width: 300px;
-  height: 350px;
-`;
-
-const TextImg = styled.img`
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  margin-bottom: 30px;
-  border-radius: 15px;
-`;
-
-const Info = styled.div`
-  color: #fff;
-  position: absolute;
-  border-radius: 15px;
-  left: 0;
-  bottom: 0;
-  background: rgba(0, 0, 0, 0.5);
-  width: 100%;
-  height: 100%;
-  padding: 15px;
-  box-sizing: border-box;
-  opacity: 0;
-  transition: opacity 0.35s ease-in-out;
-
-  ${Box}:hover & {
-    opacity: 1;
-  }
-`;
-
-const TopContainer = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-`;
-
-const IconWithCount = styled.div`
-  display: flex;
-  align-items: center;
-  margin-right: 10px;
-`;
-
-const Count = styled.div`
-  margin-left: 5px;
-`;
-
-const Heart = styled(BsFillSuitHeartFill)`
-  width: 30px;
-  height: 30px;
-`;
-
-const LookUp = styled(BsEyeFill)`
-  width: 30px;
-  height: 30px;
-`;
-
-const TitleText = styled.h2`
-  font-size: 30px;
-`;
-
-const ScheduleeText = styled.h5`
-  font-size: 20px;
-  font-weight: 500;
-`;
-
-const DateText = styled.p`
-  position: absolute;
-  bottom: 15px;
-  right: 15px;
 `;
